Add logo to sidebar and slightly modify light theme
This is achieved by adding one line to the index.hbs. This is not ideal
as we fix the whole theme by it and we should make sure to keep it up to
date with upstream (using `mdbook init --theme`).

Modifications to index.hbs are indicated by <!-- MODIFICATION START -->
and <!-- MODIFICATION END -->.

There is this relevant issue upstream: https://github.com/rust-lang/mdBook/pull/1584